types: improve memfs safety
